elf: Move functions only used in tests
This is a follow-up to #18326. A few functions are only used in tests,
so golangci-lint complains about them not being used (it's not aware of
the build tags).

Fix this by moving the functions into the test file.
